fix: resolve all P1/P2 findings from third review
- Fix Tauri commands.rs u32→u64 migration (handle types, mdns browse/advertise,
  alloc_fetch_token arg, session_max_datagram_size cast)
- Drop phantom compression_level from Tauri CompressionOptions construction
- Fix Deno session function BigInt conversions: all call() return values for handle
  types now explicitly wrapped in BigInt() (sessionConnect, createBidiStream,
  nextBidiStream, createUniStream, nextUniStream)
- Fix Deno call() type annotations to use number (JSON reality) with BigInt() at boundary
- All 48/48 integration tests pass; workspace and Tauri cargo check clean

+++ packages/iroh-http-tauri/src/state.rs
@@ -15,15 +15,15 @@ fn endpoint_slab() -> &'static Mutex<Slab<IrohEndpoint>> {
     S.get_or_init(|| Mutex::new(Slab::new()))
 }
 
-pub fn insert_endpoint(ep: IrohEndpoint) -> u32 {
-    endpoint_slab().lock().unwrap().insert(ep) as u32
+pub fn insert_endpoint(ep: IrohEndpoint) -> u64 {
+    endpoint_slab().lock().unwrap().insert(ep) as u64
 }
 
-pub fn get_endpoint(handle: u32) -> Option<IrohEndpoint> {
+pub fn get_endpoint(handle: u64) -> Option<IrohEndpoint> {
     endpoint_slab().lock().unwrap().get(handle as usize).cloned()
 }
 
-pub fn remove_endpoint(handle: u32) -> Option<IrohEndpoint> {
+pub fn remove_endpoint(handle: u64) -> Option<IrohEndpoint> {
     let mut slab = endpoint_slab().lock().unwrap();
     if slab.contains(handle as usize) {
         Some(slab.remove(handle as usize))
@@ -36,7 +36,7 @@ pub fn remove_endpoint(handle: u32) -> Option<IrohEndpoint> {
 // (Currently unused — commands.rs uses iroh_http_core::stream directly.)
 
 /// Allocate a writer handle and stash the paired reader so rawFetch can claim it.
-pub fn js_alloc_body_writer() -> u32 {
+pub fn js_alloc_body_writer() -> u64 {
     let (handle, reader) = alloc_body_writer();
     store_pending_reader(handle, reader);
     handle
